Get flash player runner compiling
